Move package into download directory, move tests into tests
