+++ minis/downloader/downloadAllMinis.py
