+++ minis/setup.py
@@ -1,0 +1,3 @@
+from setuptools import setup, find_packages
+
+setup(name="downloader", packages=find_packages())
